Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/083-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'top.log')
FAILED (This TC requires JavaScript enabled)

Harness Error (FAIL), message = TypeError: null is not an object (evaluating 'top.log')

FAIL  scheduler: event listener defined by script in a document in history assert_array_equals: lengths differ, expected array ["inline script #1", "IFRAME script", "end script #1", "event: foo", "inline script #2"] length 5, got ["inline script #1", "IFRAME script", "end script #1", "event: foo", "event: foo", "inline script #2"] length 6
